Migrate stories for PostPreview to Doctocat #507
