+++ .changeset/brown-ducks-sit.md
@@ -1,0 +1,6 @@
+---
+'@undataforum/components': minor
+---
+
+Remove top margin from button in PostPreview and enforce top down margins
+(bottom margins)
