+++ .changeset/nasty-donuts-rhyme.md
@@ -1,0 +1,5 @@
+---
+'@undataforum/preset': minor
+---
+
+Add variants for Text
