+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -9,5 +9,7 @@
       url: /components/event-preview
     - title: GitHubIcon
       url: /components/github-icon
+    - title: PostPreview
+      url: /components/post-preview
     - title: TwitterIcon
       url: /components/twitter-icon
