+++ docs/src/pages/event-preview/grid-of-events.js
@@ -7,7 +7,7 @@ const events = [
     id: '0e0dde9c-1cb3-4bbf-a070-d3b838d98398',
     tag: 'Webinar',
     title: (
-      <Heading as="h1" sx={{ textAlign: 'start', mb: 3 }}>
+      <Heading as="h2" mb={3}>
         Ut aut reiciendis voluptatibus maiores alias consequatur
       </Heading>
     ),
@@ -25,7 +25,7 @@ const events = [
     date: 'Display date',
     duration: 'display duration',
     description: (
-      <Text as="p" sx={{ lineHeight: 'body', textAlign: 'start', mb: 3 }}>
+      <Text as="p" variant="text.default" mb={3}>
         Totam rem aperiam. Inventore veritatis et quasi architecto beatae vitae
         dicta sunt explicabo. Lorem ipsum dolor sit amet, consectetur
         adipisicing elit.
@@ -38,7 +38,7 @@ const events = [
     id: 'c539b355-4729-4d8f-88eb-ee955e1ae832',
     tag: 'Session',
     title: (
-      <Heading as="h1" sx={{ textAlign: 'start', mb: 3 }}>
+      <Heading as="h2" mb={3}>
         Do eiusmod tempor incididunt ut labore et dolore magna aliqua
       </Heading>
     ),
@@ -46,7 +46,7 @@ const events = [
     date: 'Display date',
     duration: 'display duration',
     description: (
-      <Text as="p" sx={{ lineHeight: 'body', textAlign: 'start', mb: 3 }}>
+      <Text as="p" variant="text.default" mb={3}>
         Laboris nisi ut aliquip ex ea commodo consequat. Ut enim ad minim
         veniam, quis nostrud exercitation ullamco. Lorem ipsum dolor sit amet,
         consectetur adipisicing elit. Do eiusmod tempor incididunt ut labore et
@@ -62,7 +62,7 @@ const events = [
     id: '60b7bd4e-b6d1-4f49-8f0a-4aaff9e55015',
     tag: 'Workshop',
     title: (
-      <Heading as="h1" sx={{ textAlign: 'start', mt: 1, mb: 3 }}>
+      <Heading as="h2" mb={3}>
         Et harum quidem rerum facilis est et expedita distinctio
       </Heading>
     ),
@@ -70,7 +70,7 @@ const events = [
     date: 'Display date',
     duration: 'display duration',
     description: (
-      <Text as="p" sx={{ lineHeight: 'body', textAlign: 'start', mb: 3 }}>
+      <Text as="p" variant="text.default" mb={3}>
         Corrupti quos dolores et quas molestias excepturi sint occaecati. Animi,
         id est laborum et dolorum fuga. Temporibus autem quibusdam et aut
         officiis debitis aut rerum necessitatibus saepe eveniet ut et voluptates
@@ -84,7 +84,7 @@ const events = [
     id: 'e87d27fd-53e5-438a-93c0-32ea28e91f1a',
     tag: 'Panel',
     title: (
-      <Heading as="h1" sx={{ textAlign: 'start', mt: 1, mb: 3 }}>
+      <Heading as="h2" mb={3}>
         Itaque earum rerum hic tenetur a sapiente delectus
       </Heading>
     ),
@@ -97,7 +97,7 @@ const events = [
     date: 'Display date',
     duration: 'display duration',
     description: (
-      <Text as="p" sx={{ lineHeight: 'body', textAlign: 'start', mb: 3 }}>
+      <Text as="p" variant="text.default" mb={3}>
         Nisi ut aliquid ex ea commodi consequatur? Facere possimus, omnis
         voluptas assumenda est, omnis dolor repellendus. Quis autem vel eum iure
         reprehenderit qui in ea voluptate velit esse quam.
